update index.html, app.js, and README.md
I've made some changes to these three files.

The first (index.html), I recently discovered adding custom attributes
to HTML elements and found a particular use for them. I added to the
numerical buttons a custom attribute called "data-button-type" and will
do the same for the other buttons. This custom attribute is intended to
help my app.js file to identify the type of button so it can call the
appropriate function. The other change made to index.html was I moved
the <script> element to the <head> section and gave it the "defer"
attribute so that the browser loads the script after the HTML has
fully loaded.

The second, (app.js), I've deleted a lot of code and rewritten them
after finding better, more efficient alternatives. I've updated the
addEventListener() method's callback function to be a more succinct
arrow function and have changed the code in the event handler arrow
function which uses my custom data attributes.

The last (README.md), I've added a "Features Used" section to the
README file which will mention the techniques I've used.